requirements.txt
pyramid==2.0.2
pyramid-jinja2==2.10
pyramid-debugtoolbar==4.12.1
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
marshmallow==3.21.0
waitress==2.1.2
python-dotenv==1.0.0
werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
webtest==3.0.0